+++ docs/conf.py
@@ -62,5 +62,4 @@ numpydoc_xref_param_type = True
 
 intersphinx_mapping = {
     "python": ("https://docs.python.org/3", None),
-    "requests": ("https://requests.readthedocs.io/en/latest/", None),
 }